+++ src/main/java/com/eni/pizzaWebsite/bll/OrderManager.java
@@ -14,7 +14,7 @@ public class OrderManager implements IOrderManager{
     }
 
     @Override
-    public void addProductToOrder(Product product, Long idClient,int quantity ) {
+    public void addProductToOrder(Product product, Long idClient,Long quantity ) {
         orderDao.addProductToOrder(product, idClient);
     }
 }
